Propagate status when editing most recent update.
When editing the most recent update, a change in status is likely to indicate a
change in the overall status of the incident (such as changing the update from
`Identified` to `Resolved`). As such, when the most recent update is edited, we
are going to propagate the selected status to both the overall incident and the
new update form.

Closes #5