Include Date in the Started field of Slack message.
While the Slack message is geared towards readability during an active
incident, it can be helpful to see which _date_ the incident started on. This
way, as people are scrolling up through Slack, they don't have to proceed past
an incident - to the preceding date horizontal-rule - in order to figure out on
what date the incident occurred.

Closes #8