Prepare for flux-inspired architecture.
Right now, the state management in this app is kind of crazy. There are lots of
places where we duplicate logic for updating output values based on actions in
the application. I'd like to move the application to a more Flux-inspired state
management system. But, before I get to that, I want to reorganize the file
structure so it's not just one giant bag of files. Should should make it easier
to see where the state management files will go.