Ensure timezone is parsed an abbreviation.
The `RegExp` that was parsing the `date.toTimeString()` value was expecting the
timezone substring to always contain an abbreviation (such as `EDT`). However,
on some computers, the substring is a longer value like, `Eastern Daylight Time`.
Such values were breaking the Time pipe.

This update increases the flexibility of the Time pipe in order to account for
both long and short version; and then, converts the long version to an
abbreviation when necessary.

Closes #14